Tiding up Serialization Tests.
