+++ src/Tester/SerializationTests/ConfigurationTests/ClientConfigurationForSerializer.xml
