+++ src/Tester/SerializationTests/ConfigurationTests/ClientConfigurationForSerializer2.xml
