+++ src/Tester/SerializationTests/ConfigurationTests/ClientConfigurationForSerializer3.xml
