+++ src/Tester/SerializationTests/ConfigurationTests/ClientConfigurationForSerializer4.xml
