+++ src/Tester/SerializationTests/ConfigurationTests/ClientConfigurationForSerializer5.xml
