+++ src/Tester/SerializationTests/ConfigurationTests/ClientConfigurationForSerializer6.xml
