+++ src/Tester/SerializationTests/ConfigurationTests/ClientConfigurationForSerializer7.xml
