+++ src/Tester/SerializationTests/ConfigurationTests/ClientConfigurationForSerializer8.xml
