+++ src/TesterInternal/Serialization/BuiltInSerializerTests.cs
@@ -7,8 +7,6 @@ using System.Runtime.Serialization;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Orleans;
-using Orleans.Concurrency;
 using Orleans.Messaging;
 using Orleans.Runtime;
 using Orleans.Runtime.Configuration;
@@ -17,7 +15,7 @@ using Orleans.CodeGeneration;
 
 // ReSharper disable NotAccessedVariable
 
-namespace UnitTests.SerializerTests
+namespace UnitTests.Serialization
 {
     /// <summary>
     /// Test the built-in serializers
@@ -828,7 +826,7 @@ namespace UnitTests.SerializerTests
         }
 
         [TestMethod, TestCategory("Functional"), TestCategory("Serialization")]
-        public void ValidateBuildSegmentListWithLengthLimit()
+        public void Serialize_ValidateBuildSegmentListWithLengthLimit()
         {
             byte[] array1 = { 1 };
             byte[] array2 = { 2, 3 };
@@ -1006,7 +1004,7 @@ namespace UnitTests.SerializerTests
         }
 
         [TestMethod, TestCategory("Functional"), TestCategory("Serialization")]
-        public void IsOrleansShallowCopyable()
+        public void SerializationTests_IsOrleansShallowCopyable()
         {
             Type t = typeof(Dictionary<string, object>);
             Assert.IsFalse(t.IsOrleansShallowCopyable(), "IsOrleansShallowCopyable: {0}", t.Name);
