+++ src/TesterInternal/TesterInternal.csproj
@@ -89,10 +89,10 @@
     <Reference Include="System.XML" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="General\BuiltInSerializerTests.cs" />
+    <Compile Include="Serialization\BuiltInSerializerTests.cs" />
     <Compile Include="General\Identifiertests.cs" />
     <Compile Include="General\InterfaceRules.cs" />
-    <Compile Include="General\FallbackBuiltInSerializationTests.cs" />
+    <Compile Include="Serialization\FallbackBuiltInSerializationTests.cs" />
     <Compile Include="MembershipTests\MySqlMembershipTableTests.cs" />
     <Compile Include="OrleansRuntime\AsyncSerialExecutorTests.cs" />
     <Compile Include="OrleansRuntime\ExceptionsTests.cs" />
